rename util -> schema-loader-utils, let -> const
